integrations/chainlit/chainlit_simple_chat/requirement.txt
chainlit
judini
openai
python-dotenv